+++ src/service/mod.rs
@@ -8,3 +8,4 @@
 pub mod preview_service;
 pub mod repo_service;
 pub mod status_service;
+pub mod zip_service;